feat: Move the Rust repo to the right place
