--- Rust/Cargo.toml
@@ -1,7 +1,0 @@
-[package]
-name = "MyBot"
-version = "0.1.0"
-authors = ["Kerollmops <renault.cle@gmail.com>",
-           "Paul Butler <paul.butler@twosigma.com>"]
-
-[dependencies]
