--- Rust/run_game.bat
@@ -1,2 +1,0 @@
-cargo rustc --release -q -- -Awarnings
-.\halite -d "240 160" "target\release\MyBot" "target\release\MyBot"
