--- Rust/src/hlt/mod.rs
@@ -1,9 +1,0 @@
-pub mod entity;
-pub mod game;
-pub mod logging;
-pub mod command;
-pub mod constants;
-pub mod game_map;
-pub mod player;
-pub mod collision;
-mod parse;
