+++ airesources/Rust/Cargo.toml
@@ -1,10 +1,7 @@
 [package]
-name = "halite_player"
+name = "MyBot"
 version = "0.1.0"
-authors = ["Paul Butler <paul.butler@twosigma.com>"]
+authors = ["Kerollmops <renault.cle@gmail.com>",
+           "Paul Butler <paul.butler@twosigma.com>"]
 
 [dependencies]
-
-[[bin]]
-name = "MyBot"
-path = "src/main.rs"
