+++ airesources/Rust/src/hlt/entity/docking_status.rs
