+++ airesources/Rust/src/hlt/entity/ship.rs
