+++ airesources/Rust/src/hlt/game.rs
@@ -1,5 +1,4 @@
-
-use std::io::stdin;
+use std::io::{stdin, stdout, Write};
 use hlt::parse::Decodable;
 use hlt::entity::GameState;
 use hlt::command::Command;
@@ -16,14 +15,14 @@ impl Game {
     fn read_line() -> String {
         let mut buffer = String::new();
         stdin().read_line(&mut buffer).expect("Read error");
-        return buffer;
+        buffer
     }
 
     fn read_id() -> usize {
         let line = Game::read_line();
         let parts = line.split_whitespace();
         let mut iter = parts.into_iter();
-        return usize::parse(&mut iter);
+        usize::parse(&mut iter)
     }
 
     fn read_size() -> (i32, i32) {
@@ -32,7 +31,7 @@ impl Game {
         let mut iter = parts.into_iter();
         let width = i32::parse(&mut iter);
         let height = i32::parse(&mut iter);
-        return (width, height);
+        (width, height)
     }
 
     pub fn new(name: &str) -> Game {
@@ -41,13 +40,14 @@ impl Game {
 
         println!("{}", name);
 
-        let game = Game {
+        // ignore first line
+        Game::read_line();
+
+        Game {
             my_id,
             map_width,
             map_height,
-        };
-        game.update_map();
-        game
+        }
     }
 
     pub fn update_map(&self) -> GameMap {
@@ -55,13 +55,14 @@ impl Game {
         let parts = line.split_whitespace();
         let mut iter = parts.into_iter();
         let game_state = GameState::parse(&mut iter);
-        return GameMap::new(self, game_state);
+        GameMap::new(self, game_state)
     }
 
-    pub fn send_command_queue(&self, commands: Vec<Command>) {
+    pub fn send_command_queue(&self, commands: &[Command]) {
         for command in commands {
-            print!("{}", command.encode());
+            let encoded = command.encode();
+            stdout().write(encoded.as_bytes()).unwrap();
         }
-        println!();
+        println!()
     }
 }
