+++ airesources/Rust/src/hlt/game_map.rs
@@ -1,33 +1,39 @@
-
 use hlt::game::Game;
 use hlt::entity::{GameState, Planet};
 use hlt::player::Player;
 use hlt::collision::intersect_segment_circle;
 use hlt::entity::{Entity, Ship};
 
+/// Map which houses the current game information/metadata.
 pub struct GameMap<'a> {
     game: &'a Game,
     state: GameState,
 }
 
 impl<'a> GameMap<'a> {
-    pub fn new(game: &Game, state: GameState) -> GameMap {
-        return GameMap { game, state };
+    pub fn new(game: &'a Game, state: GameState) -> Self {
+        Self { game, state }
     }
 
-    pub fn all_planets(&self) -> &Vec<Planet> {
+    /// Return your own player.
+    pub fn me(&self) -> &Player {
+        let my_id = self.game.my_id;
+        &self.state.players[my_id]
+    }
+
+    /// Returns all planets at the actual game state.
+    pub fn all_planets(&self) -> &[Planet] {
         &self.state.planets
     }
 
-    pub fn get_me(&self) -> &Player {
-        let my_id = self.game.my_id;
-        let player = &self.state.players[my_id];
-        return player;
+    /// Returns all players at the actual game state including yourself.
+    pub fn all_players(&self) -> &[Player] {
+        &self.state.players
     }
 
     pub fn obstacles_between<T: Entity>(&self, ship: &Ship, target: &T) -> bool {
         for planet in self.all_planets() {
-            if intersect_segment_circle(ship, target, planet, ship.get_radius() + 0.1) {
+            if intersect_segment_circle(ship, target, planet, ship.radius() + 0.1) {
                 return true;
             }
         }
