+++ airesources/Rust/src/hlt/parse.rs
@@ -1,4 +1,3 @@
-
 use std::str::FromStr;
 use std::fmt::Debug;
 
@@ -13,13 +12,12 @@ where
     I: Iterator<Item = &'a str>,
     <T as FromStr>::Err: Debug,
 {
-    let next_token = &tokens.next().expect(&"Expected token, got EOF");
-    let value: T = next_token.parse().expect(&format!(
+    let next_token = tokens.next().expect("Expected token, got EOF");
+    next_token.parse::<T>().unwrap_or_else(|_| panic!(
         "Could not parse {:?} as expected type {:?}",
-        &next_token,
+        next_token,
         type_name
-    ));
-    return value;
+    ))
 }
 
 impl Decodable for f64 {
@@ -27,7 +25,7 @@ impl Decodable for f64 {
     where
         I: Iterator<Item = &'a str>,
     {
-        return parse_next_primitive(tokens, "f64");
+        parse_next_primitive(tokens, "f64")
     }
 }
 
@@ -36,7 +34,7 @@ impl Decodable for i32 {
     where
         I: Iterator<Item = &'a str>,
     {
-        return parse_next_primitive(tokens, "i32");
+        parse_next_primitive(tokens, "i32")
     }
 }
 
@@ -45,7 +43,7 @@ impl Decodable for usize {
     where
         I: Iterator<Item = &'a str>,
     {
-        return parse_next_primitive(tokens, "usize");
+        parse_next_primitive(tokens, "usize")
     }
 }
 
@@ -55,26 +53,26 @@ impl Decodable for bool {
         I: Iterator<Item = &'a str>,
     {
         let p = parse_next_primitive(tokens, "bool");
-        return match p {
+        match p {
             0 => false,
             1 => true,
-            _ => panic!(format!("Expected 0 or 1, got {:?}", p)),
-        };
+            _ => panic!("Expected 0 or 1, got {:?}", p),
+        }
     }
 }
 
-impl<T: Decodable> Decodable for Vec<T> {
-    fn parse<'a, I>(tokens: &mut I) -> Vec<T>
+impl<T: Decodable> Decodable for Box<[T]> {
+    fn parse<'a, I>(tokens: &mut I) -> Box<[T]>
     where
         I: Iterator<Item = &'a str>,
     {
-        let size = parse_next_primitive(tokens, "i64");
-        let mut result = Vec::with_capacity(size);
+        let size: i64 = parse_next_primitive(tokens, "i64");
+        let mut result = Vec::with_capacity(size as usize);
         for _ in 0..size {
             let p = T::parse(tokens);
             result.push(p);
         }
-        return result;
+        result.into_boxed_slice()
     }
 }
 
@@ -83,13 +81,12 @@ impl<T: Decodable> Decodable for Option<T> {
     where
         I: Iterator<Item = &'a str>,
     {
-
         let present = bool::parse(tokens);
         let value = T::parse(tokens);
 
-        return match present {
+        match present {
             true => Some(value),
             false => None,
-        };
+        }
     }
 }
