+++ airesources/Rust/src/main.rs
@@ -9,23 +9,24 @@ mod hlt;
 use hlt::entity::{Entity, DockingStatus};
 use hlt::game::Game;
 use hlt::logging::Logger;
-use hlt::command::Command;
 
 fn main() {
+    let name = "Kevina";
     // Initiailize the game
-    let game = Game::new("Settler");
+    let game = Game::new(name);
     // Initialize logging
     let mut logger = Logger::new(game.my_id);
-    logger.log("Starting my Settler bot!");
+    logger.log(&format!("Starting my {} bot!", name));
 
-    // For each turn
+    let mut command_queue = Vec::new();
+
     loop {
+
         // Update the game state
         let game_map = game.update_map();
-        let mut command_queue: Vec<Command> = Vec::new();
 
         // Loop over all of our player's ships
-        for ship in game_map.get_me().all_ships() {
+        for ship in game_map.me().all_ships() {
             // Ignore ships that are docked or in the process of docking
             if ship.docking_status != DockingStatus::UNDOCKED {
                 continue;
@@ -38,21 +39,22 @@ fn main() {
                     continue;
                 }
 
+                // If we are close enough to dock, do it!
                 if ship.can_dock(planet) {
-                    // If we are close enough to dock, do it!
                     command_queue.push(ship.dock(planet))
                 } else {
                     // If not, navigate towards the planet
                     let navigate_command = ship.navigate(&ship.closest_point_to(planet, 3.0), &game_map, 90);
-                    match navigate_command {
-                        Some(command) => command_queue.push(command),
-                        _ => {}
+                    if let Some(command) = navigate_command {
+                        command_queue.push(command)
                     }
+
                 }
                 break;
             }
         }
         // Send our commands to the game
-        game.send_command_queue(command_queue);
+        game.send_command_queue(&command_queue);
+        command_queue.clear();
     }
 }
